requirements.txt
flask>=3.0.2
pycryptodome>=3.20.0
pyjwt[crypto]>=2.8.0